requirements.txt
Django==3.2.10
djangorestframework==3.12.4
mysqlclient==2.0.3